Update VehicleMVCController - implemented creation of vehicle;

+++ src/main/java/com/example/smartgarage/models/dtos/VehicleFilterDto.java
@@ -19,7 +19,7 @@ public class VehicleFilterDto {
     private String brand;
     private String model;
     private User user;
-    private int creationYear;
+    private Long creationYear;
     private String sortBy;
     private String sortOrder;
 }
